restructure files under package
