--- README.md
@@ -1,3 +1,0 @@
-# NICEXAUUSD
-
-Automated trading research environment for gold (XAUUSD). Includes Sniper strategy with fallback configs.
