+++ nicegold_v5/README.md
@@ -1,0 +1,3 @@
+# NICEXAUUSD
+
+Automated trading research environment for gold (XAUUSD). Includes Sniper strategy with fallback configs.
