--- tests/conftest.py
@@ -1,6 +1,0 @@
-import sys
-from pathlib import Path
-
-# Add project root to sys.path
-ROOT = Path(__file__).resolve().parents[1]
-sys.path.append(str(ROOT))
